- Clean up code - Add Localisation
