+++ android/app/build.gradle
@@ -54,7 +54,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.nartus.interactiveDiary.dev"
+        applicationId "com.nartus.interactiveDiary"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-build-configuration.
         minSdkVersion 29
