+++ test/mock_firebase.dart
@@ -16,9 +16,3 @@ void setupFirebaseAuthMocks([Callback? customHandlers]) {
   setupFirebaseCoreMocks();
 }
 
-Future<T> neverEndingFuture<T>() async {
-  // ignore: literal_only_boolean_expressions
-  while (true) {
-    await Future.delayed(const Duration(minutes: 5));
-  }
-}
